Remove RenderPassSubpassInterface and RenderPassCompatible traits
