+++ vulkano/src/framebuffer/mod.rs
@@ -102,8 +102,6 @@ pub use self::sys::RenderPassCreationError;
 pub use self::sys::RenderPassDescReal;
 pub use self::sys::RenderPassSys;
 pub use self::traits::FramebufferAbstract;
-pub use self::traits::RenderPassCompatible;
-pub use self::traits::RenderPassSubpassInterface;
 pub use self::traits::Subpass;
 
 #[macro_use]
